refactor(client): P8(1/2) — autonomous loop onto ToolLoopRunner, duplicate deleted
_run_loop_iteration (308 lines) moves verbatim into ToolLoopRunner.run_
autonomous (host-based); _dispatch_loop_tool(+inner) become runner
methods; _LoopMessageProxy/_LoopAuthorProxy move to tool_loop.py with a
client re-export alias (tests import them from client). Bot keeps
delegates for all three (api.py calls _run_loop_iteration; the scheduled
path and tests call the dispatch pair).

Both pipelines now live in one module sharing one dispatch table. Every
sec-4.3 asymmetry pin green: CircuitOpenError re-raise, no guards/
classifier on the loop path, raw llm_client seam (no gateway wiring),
cap -> is_error/failure_class=cancelled with stale-partial surfacing,
recovered-error reflection semantics, export staging, proxy dispatch.
